requirements.txt
pycurl==7.19.3.1
altgraph==0.10.2
bdist-mpkg==0.5.0
dj-database-url==0.4.2
reportlab==3.4.0
pillow>=2.4.0
olefile=0.44
Django==1.11.6
django-cors-headers==2.1.0
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
macholib==1.5.1
modulegraph==0.10.4
py2app==0.7.3
PyJWT==1.5.3
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
whitenoise==3.3.1
zope.interface==4.1.1
psycopg2==2.5.4
